apps/api/pyproject.toml
[project]
name = "lecpa-api"
version = "0.1.0"
description = "FastAPI backend for Krystal Le Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pgvector>=0.2.4",
    "boto3>=1.34.0",
    "anthropic>=0.18.0",
    "sentence-transformers>=2.3.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "orjson>=3.9.0",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["database", "routers", "services"]
